Title: 합을 옆으로 밀어 출력, Time: 102ms, Memory: 9MB, Status: Passed - CodeTree

+++ 240827/합을 옆으로 밀어 출력/push-the-sum-sideways-to-output.kt
@@ -10,7 +10,7 @@ class Main {
     private var sum = 0
 
     fun solve() {
-        n = br.readLine()
+        n = br.readLine().toInt()
 
         repeat(n) {
             sum += br.readLine().toInt()
